requirements.txt
# Dépendances principales pipeline nouvelle architecture
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
rapidfuzz==3.9.6
unidecode==1.3.8
tqdm==4.66.5
# Optionnel pour graphiques/analyses avancées
pandas==2.2.2
matplotlib==3.9.2
seaborn==0.13.2
# Tests
pytest==8.3.3
pytest-cov==5.0.0
